move table query to api directory
