+++ packages/psss/src/containers/Tables/CountriesTable.js
@@ -8,7 +8,7 @@ import { ExpandableTable, ExpandableTableBody } from '@tupaia/ui-components';
 import { connect } from 'react-redux';
 import { COLUMN_WIDTHS } from './constants';
 import { CountrySummaryTable } from './CountrySummaryTable';
-import { useTableQuery } from '../../hooks';
+import { useTableQuery } from '../../api';
 import {
   createTotalCasesAccessor,
   AlertCell,
